requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pymysql
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
alembic
pydantic_settings
pillow 
python-multipart